More fixes

+++ numba/tests/test_dyn_array.py
@@ -1112,7 +1112,7 @@ class TestNpArray(MemoryLeakMixin, BaseTest):
                            "('b', '<f4')])) to a homogenous type")):
             st = np.dtype([('a', 'i4'), ('b', 'f4')])
             val = np.zeros(1, dtype=st)[0]
-            cfunc(((1, 2), (1, val)))
+            cfunc(((1, 2), (np.int64(1), val)))
 
 
 class TestNpConcatenate(MemoryLeakMixin, TestCase):
